NMSIS/DSP: Sync changes for DSP library from 5.9.0

+++ NMSIS/DSP/Source/MatrixFunctions/riscv_mat_add_f16.c
@@ -96,13 +96,13 @@ riscv_status riscv_mat_add_f16(
       /* C(m,n) = A(m,n) + B(m,n) */
 
       /* Add and store result in destination buffer. */
-      *pOut++ = *pInA++ + *pInB++;
+      *pOut++ = (_Float16)*pInA++ + (_Float16)*pInB++;
 
-      *pOut++ = *pInA++ + *pInB++;
+      *pOut++ = (_Float16)*pInA++ + (_Float16)*pInB++;
 
-      *pOut++ = *pInA++ + *pInB++;
+      *pOut++ = (_Float16)*pInA++ + (_Float16)*pInB++;
 
-      *pOut++ = *pInA++ + *pInB++;
+      *pOut++ = (_Float16)*pInA++ + (_Float16)*pInB++;
 
       /* Decrement loop counter */
       blkCnt--;
@@ -123,7 +123,7 @@ riscv_status riscv_mat_add_f16(
       /* C(m,n) = A(m,n) + B(m,n) */
 
       /* Add and store result in destination buffer. */
-      *pOut++ = *pInA++ + *pInB++;
+      *pOut++ = (_Float16)*pInA++ + (_Float16)*pInB++;
 
       /* Decrement loop counter */
       blkCnt--;
