+++ NMSIS/DSP/Source/MatrixFunctions/riscv_mat_sub_f16.c
@@ -94,10 +94,10 @@ riscv_status riscv_mat_sub_f16(
       /* C(m,n) = A(m,n) - B(m,n) */
 
       /* Subtract and store result in destination buffer. */
-      *pOut++ = (*pInA++) - (*pInB++);
-      *pOut++ = (*pInA++) - (*pInB++);
-      *pOut++ = (*pInA++) - (*pInB++);
-      *pOut++ = (*pInA++) - (*pInB++);
+      *pOut++ = (_Float16)(*pInA++) - (_Float16)(*pInB++);
+      *pOut++ = (_Float16)(*pInA++) - (_Float16)(*pInB++);
+      *pOut++ = (_Float16)(*pInA++) - (_Float16)(*pInB++);
+      *pOut++ = (_Float16)(*pInA++) - (_Float16)(*pInB++);
 
       /* Decrement loop counter */
       blkCnt--;
@@ -118,7 +118,7 @@ riscv_status riscv_mat_sub_f16(
       /* C(m,n) = A(m,n) - B(m,n) */
 
       /* Subtract and store result in destination buffer. */
-      *pOut++ = (*pInA++) - (*pInB++);
+      *pOut++ = (_Float16)(*pInA++) - (_Float16)(*pInB++);
 
       /* Decrement loop counter */
       blkCnt--;
