+++ NMSIS/DSP/Source/StatisticsFunctions/riscv_absmax_f16.c
@@ -66,7 +66,7 @@ void riscv_absmax_f16(
   outIndex = 0U;                                                                                            \
   /* Load first input value that act as reference value for comparision */                                  \
   out = *pSrc++;                                                                                            \
-  out = (out > 0.0f16) ? out : -out;                                                                             \
+  out = ((_Float16)out > 0.0f16) ? out : -(_Float16)out;                                                                             \
   /* Initialize index of extrema value. */                                                                  \
   index = 0U;                                                                                               \
                                                                                                             \
@@ -77,9 +77,9 @@ void riscv_absmax_f16(
   {                                                                                                         \
     /* Initialize cur_absmax to next consecutive values one by one */                                         \
     cur_absmax = *pSrc++;                                                                                     \
-    cur_absmax = (cur_absmax > 0.0f16) ? cur_absmax : -cur_absmax;                                                                 \
+    cur_absmax = ((_Float16)cur_absmax > 0.0f16) ? cur_absmax : -(_Float16)cur_absmax;                                                                 \
     /* compare for the extrema value */                                                                     \
-    if (cur_absmax > out)                                                                         \
+    if ((_Float16)cur_absmax > (_Float16)out)                                                                         \
     {                                                                                                       \
       /* Update the extrema value and it's index */                                                         \
       out = cur_absmax;                                                                                       \
@@ -87,24 +87,24 @@ void riscv_absmax_f16(
     }                                                                                                       \
                                                                                                             \
     cur_absmax = *pSrc++;                                                                                     \
-    cur_absmax = (cur_absmax > 0.0f16) ? cur_absmax : -cur_absmax;                                                                 \
-    if (cur_absmax > out)                                                                         \
+    cur_absmax = ((_Float16)cur_absmax > 0.0f16) ? cur_absmax : -(_Float16)cur_absmax;                                                                 \
+    if ((_Float16)cur_absmax > (_Float16)out)                                                                         \
     {                                                                                                       \
       out = cur_absmax;                                                                                       \
       outIndex = index + 2U;                                                                                \
     }                                                                                                       \
                                                                                                             \
     cur_absmax = *pSrc++;                                                                                     \
-    cur_absmax = (cur_absmax > 0.0f16) ? cur_absmax : -cur_absmax;                                                                 \
-    if (cur_absmax > out)                                                                          \
+    cur_absmax = ((_Float16)cur_absmax > 0.0f16) ? cur_absmax : -(_Float16)cur_absmax;                                                                 \
+    if ((_Float16)cur_absmax > (_Float16)out)                                                                          \
     {                                                                                                       \
       out = cur_absmax;                                                                                       \
       outIndex = index + 3U;                                                                                \
     }                                                                                                       \
                                                                                                             \
     cur_absmax = *pSrc++;                                                                                     \
-    cur_absmax = (cur_absmax > 0.0f16) ? cur_absmax : -cur_absmax;                                                                 \
-    if (cur_absmax > out)                                                                          \
+    cur_absmax = ((_Float16)cur_absmax > 0.0f16) ? cur_absmax : -(_Float16)cur_absmax;                                                                 \
+    if ((_Float16)cur_absmax > (_Float16)out)                                                                          \
     {                                                                                                       \
       out = cur_absmax;                                                                                       \
       outIndex = index + 4U;                                                                                \
@@ -123,8 +123,8 @@ void riscv_absmax_f16(
   while (blkCnt > 0U)                                                                                       \
   {                                                                                                         \
     cur_absmax = *pSrc++;                                                                                     \
-    cur_absmax = (cur_absmax > 0.0f16) ? cur_absmax : -cur_absmax;                                                                 \
-    if (cur_absmax > out)                                                                         \
+    cur_absmax = ((_Float16)cur_absmax > 0.0f16) ? cur_absmax : -(_Float16)cur_absmax;                                                                 \
+    if ((_Float16)cur_absmax > (_Float16)out)                                                                         \
     {                                                                                                       \
       out = cur_absmax;                                                                                       \
       outIndex = blockSize - blkCnt;                                                                        \
@@ -136,7 +136,7 @@ void riscv_absmax_f16(
                                                                                                             \
   /* Store the extrema value and it's index into destination pointers */                                    \
   *pResult = out;                                                                                           \
-  *pIndex = outIndex;  
+  *pIndex = outIndex;
 }
 #else
 void riscv_absmax_f16(
@@ -152,7 +152,7 @@ void riscv_absmax_f16(
   outIndex = 0U;
 
   /* Load first input value that act as reference value for comparision */
-  out = fabsf(*pSrc++);
+  out = (_Float16)fabsf((float32_t)*pSrc++);
 
   /* Initialize blkCnt with number of samples */
   blkCnt = (blockSize - 1U);
@@ -160,10 +160,10 @@ void riscv_absmax_f16(
   while (blkCnt > 0U)
   {
     /* Initialize maxVal to the next consecutive values one by one */
-    maxVal = fabsf(*pSrc++);
+    maxVal = (_Float16)fabsf((float32_t)*pSrc++);
 
     /* compare for the maximum value */
-    if (out < maxVal)
+    if ((_Float16)out < (_Float16)maxVal)
     {
       /* Update the maximum value and it's index */
       out = maxVal;
@@ -183,5 +183,5 @@ void riscv_absmax_f16(
   @} end of AbsMax group
  */
 
-#endif /* #if defined(RISCV_FLOAT16_SUPPORTED) */ 
+#endif /* #if defined(RISCV_FLOAT16_SUPPORTED) */
