+++ NMSIS/DSP/Source/StatisticsFunctions/riscv_entropy_f16.c
@@ -75,7 +75,7 @@ float16_t riscv_entropy_f16(const float16_t * pSrcA,uint32_t blockSize)
     while(blkCnt > 0)
     {
        p = *pIn++;
-       accum += p * logf(p);
+       accum += p * (_Float16)logf((float32_t)p);
        
        blkCnt--;
