+++ NMSIS/DSP/Source/StatisticsFunctions/riscv_min_f16.c
@@ -85,7 +85,7 @@ void riscv_min_f16(
     minVal = *pSrc++;
 
     /* compare for the minimum value */
-    if (out > minVal)
+    if ((_Float16)out > (_Float16)minVal)
     {
       /* Update the minimum value and it's index */
       out = minVal;
@@ -93,21 +93,21 @@ void riscv_min_f16(
     }
 
     minVal = *pSrc++;
-    if (out > minVal)
+    if ((_Float16)out > (_Float16)minVal)
     {
       out = minVal;
       outIndex = index + 2U;
     }
 
     minVal = *pSrc++;
-    if (out > minVal)
+    if ((_Float16)out > (_Float16)minVal)
     {
       out = minVal;
       outIndex = index + 3U;
     }
 
     minVal = *pSrc++;
-    if (out > minVal)
+    if ((_Float16)out > (_Float16)minVal)
     {
       out = minVal;
       outIndex = index + 4U;
@@ -135,7 +135,7 @@ void riscv_min_f16(
     minVal = *pSrc++;
 
     /* compare for the minimum value */
-    if (out > minVal)
+    if ((_Float16)out > (_Float16)minVal)
     {
       /* Update the minimum value and it's index */
       out = minVal;
